Huge update
Interfaces - added code to allow to cast to Food/Dinosaur and use the modifiers in those
Dinosaur - added a foodValue to allow to give hunger when eaten
EatAction - Updated to not give fishes corpses
Fish - now extends dinosaur not actor
HealAction - updated to make use of the interfaces. It will increase the hunger value of the dinosaur by a value sent in
huntBehaviour- item removal now wokrks properly
seekbehaviour - same as above
Player - IMPORT WAS BROKEN AND UNUSED SO WE REMOVED IT
- Added plesioaiuaisur. Water based Dino, uses HuntBehaviour
Proto - added foodValue and hitPoints
Reed - commented broken import
Veloc - NEED to change back to work with interfaces!